move find_meshes to topology.py
