+++ edisgo/opf/powermodels_opf.py
@@ -4,42 +4,15 @@ import os
 import subprocess
 import sys
 
-from typing import List, Optional
+from typing import Optional
 
-import networkx as nx
 import numpy as np
 
 from edisgo.flex_opt import exceptions
 from edisgo.io.powermodels_io import from_powermodels
+from edisgo.network.topology import Topology
 
 logger = logging.getLogger(__name__)
-
-
-def find_meshes(edisgo_obj) -> Optional[List[List[int]]]:
-    """
-    Find all meshes in the grid.
-
-    Parameters
-    ----------
-    edisgo_obj : EDisGo
-        EDisGo object.
-
-    Returns
-    -------
-    Optional[List[List[int]]]
-        List of all meshes in the grid.
-        Each mesh is represented as a list of node indices.
-        If no meshes are found, None is returned.
-    """
-    meshes = nx.cycle_basis(edisgo_obj.to_graph())
-    if meshes:
-        logger.warning(
-            "Grid contains mesh(es). This might cause problems in "
-            "the power flow or optimisation."
-        )
-        return meshes
-    else:
-        return None
 
 
 def pm_optimize(
@@ -135,7 +108,7 @@ def pm_optimize(
         Default: True.
 
     """
-    find_meshes(edisgo_obj)
+    Topology.find_meshes(edisgo_obj)
     opf_dir = os.path.dirname(os.path.abspath(__file__))
     solution_dir = os.path.join(opf_dir, "opf_solutions")
     pm, hv_flex_dict = edisgo_obj.to_powermodels(
